+++ javascript/node/selenium-webdriver/lib/promise.js
@@ -911,19 +911,6 @@ class Thenable {
   catch(errback) {}
 
   /**
-   * An alias for {@link #catch()}
-   *
-   * @param {function(*): (R|IThenable<R>)} errback The
-   *     function to call if this promise is rejected. The function should
-   *     expect a single argument: the rejection reason.
-   * @return {!ManagedPromise<R>} A new promise which will be
-   *     resolved wdith the result of the invoked callback.
-   * @deprecated Use {@link #catch()} instead.
-   * @template R
-   */
-  thenCatch(errback) {}
-
-  /**
    * Registers a listener to invoke when this promise is resolved, regardless
    * of whether the promise's value was successfully computed. This function
    * is synonymous with the {@code finally} clause in a synchronous API:
@@ -960,45 +947,7 @@ class Thenable {
    * @template R
    */
   finally(callback) {}
-
-  /**
-   * Registers a listener to invoke when this promise is resolved, regardless
-   * of whether the promise's value was successfully computed. This function
-   * is synonymous with the {@code finally} clause in a synchronous API:
-   *
-   *     // Synchronous API:
-   *     try {
-   *       doSynchronousWork();
-   *     } finally {
-   *       cleanUp();
-   *     }
-   *
-   *     // Asynchronous promise API:
-   *     doAsynchronousWork().thenFinally(cleanUp);
-   *
-   * __Note:__ similar to the {@code finally} clause, if the registered
-   * callback returns a rejected promise or throws an error, it will silently
-   * replace the rejection error (if any) from this promise:
-   *
-   *     try {
-   *       throw Error('one');
-   *     } finally {
-   *       throw Error('two');  // Hides Error: one
-   *     }
-   *
-   *     promise.rejected(Error('one'))
-   *         .thenFinally(function() {
-   *           throw Error('two');  // Hides Error: one
-   *         });
-   *
-   * @param {function(): (R|IThenable<R>)} callback The function
-   *     to call when this promise is resolved.
-   * @return {!ManagedPromise<R>} A promise that will be fulfilled
-   *     with the callback result.
-   * @deprecated Use {@link #finally()} instead.
-   * @template R
-   */
-  thenFinally(callback) {}}
+}
 
 
 /**
@@ -1278,14 +1227,6 @@ class ManagedPromise {
         null, errback, 'catch', ManagedPromise.prototype.catch);
   }
 
-  /**
-   * @override
-   * @deprecated Use {@link #catch()} instead.
-   */
-  thenCatch(errback) {
-    return this.catch(errback);
-  }
-
   /** @override */
   finally(callback) {
     var error;
@@ -1301,14 +1242,6 @@ class ManagedPromise {
         throw error;
       }
     });
-  }
-
-  /**
-   * @override
-   * @deprecated Use {@link #finally()} instead.
-   */
-  thenFinally(callback) {
-    return this.finally(callback);
   }
 
   /**
@@ -1471,22 +1404,6 @@ class Deferred {
 
   /**
    * @override
-   * @deprecated Use {@code thenCatch} from the promise property directly.
-   */
-  thenCatch(opt_eb) {
-    return this.promise.thenCatch(opt_eb);
-  }
-
-  /**
-   * @override
-   * @deprecated Use {@code thenFinally} from the promise property directly.
-   */
-  thenFinally(opt_cb) {
-    return this.promise.thenFinally(opt_cb);
-  }
-
-  /**
-   * @override
    * @deprecated Use {@code finally} from the promise property directly.
    */
   finally(opt_cb) {
@@ -1494,71 +1411,6 @@ class Deferred {
   }
 }
 Thenable.addImplementation(Deferred);
-
-
-/**
- * Registers a listener to invoke when a promise is resolved, regardless of
- * whether the promise's value was successfully computed. This function is
- * synonymous with the `finally` clause in a synchronous API:
- *
- *     // Synchronous API:
- *     try {
- *       doSynchronousWork();
- *     } finally {
- *       cleanUp();
- *     }
- *
- *     // Asynchronous promise API:
- *     promise.finally(doAsynchronousWork(), cleanUp);
- *
- * __Note:__ similar to the `finally` clause, if the registered callback returns
- * a rejected promise or throws an error, that error will silently replace the
- * rejection (if any) from the initial promise:
- *
- *     try {
- *       throw Error('one');
- *     } finally {
- *       throw Error('two');  // Hides Error: one
- *     }
- *
- *     let p = promise.rejected(Error('one'));
- *     promise.finally(p, function() {
- *       throw Error('two');  // Hides Error: one
- *     });
- *
- * @param {PROMISE_TYPE} promise The thenable, promise-like object with which
- *     to register the callback.
- * @param {function(): *} callback The function to call when the promise is
- *     resolved.
- * @return {!PROMISE_TYPE} A new promise that is chained to the callback. This
- *     promise will inherit the value of the original input if the callback
- *     does not throw or return a rejected promise.
- * @template PROMISE_TYPE
- * @deprecated
- */
-function thenFinally(promise, callback) {
-  if (!isPromise(promise)) {
-    throw TypeError('first argument must be a promise-like object');
-  }
-
-  if (typeof callback !== 'function') {
-    throw TypeError('second argument must be a function');
-  }
-
-  let error;
-  let mustThrow = false;
-  return promise
-      .then(() => callback(), (e) => {
-        error = e;
-        mustThrow = true;
-        return callback();
-      })
-      .then(() => {
-        if (mustThrow) {
-          throw error;
-        }
-      });
-}
 
 
 /**
@@ -3214,7 +3066,6 @@ module.exports = {
   map: map,
   rejected: rejected,
   setDefaultFlow: setDefaultFlow,
-  thenFinally: thenFinally,
   when: when,
 
   get LONG_STACK_TRACES() { return LONG_STACK_TRACES; },