+++ javascript/node/selenium-webdriver/test/execute_script_test.js
@@ -309,21 +309,6 @@ test.suite(function(env) {
       });
     });
 
-    describe('getInnerHtml', function() {
-      test.it('returns element innerHTML using executeScript', function() {
-        let html = driver.findElement(By.css('div.request')).getInnerHtml();
-        assert(html).equalTo('GET /common/echo HTTP/1.1');
-      });
-    });
-
-    describe('getOuterHtml', function() {
-      test.it('returns element outerHTML using executeScript', function() {
-        let html = driver.findElement(By.css('div.request')).getOuterHtml();
-        assert(html)
-            .equalTo('<div class="request">GET /common/echo HTTP/1.1</div>');
-      });
-    });
-
     describe('async timeouts', function() {
       var TIMEOUT_IN_MS = 200;
       var ACCEPTABLE_WAIT = TIMEOUT_IN_MS / 10;